requirements.txt
django
djangorestframework
djangorestframework-simplejwt
requests
types-requests
black
flake8
isort
mypy
pytest
